The kernel measures its own costs as distributions
 - Measure per-core distributions for context switches, lock hold/wait, doorbell round
   trips, and IRQ-to-userspace latency.
 - Report percentiles from log-linear histograms without wrapping oversized samples.
 - Reject IRQ measurements that mix core clocks and distinguish local from cross-core
   wakeups.
 - Gate benchmarks on the required hardware authority.
 - Retry refused console lines.
 - Carry the core clock rate at 64 bits from the architecture seam to userspace.
 - Validate captures before consuming benchmark results.
 - Drive the bench rig by command and freeze one pre-optimisation baseline across
   measurable boards.

+++ arch/arm/chip/xmc4800/chip_xmc4800.cc
@@ -395,7 +395,7 @@ uint64_t arch_clock_now(void)
 // generic coherence tail (baud re-derive, SysTick re-arm) runs in cpu_clock_set. Called
 // privileged, IRQs already masked (single-core: the timer is quiesced). fCCU=fSYS and
 // fPERIPH=fCPU/2 both follow SystemCoreClock, so the CCU40 clock AND the USIC baud move.
-uint32_t arch_cpu_clock_set(uint32_t target)
+uint64_t arch_cpu_clock_set(uint32_t target)
 {
     uint32_t const previous = SystemCoreClock;
     uint32_t want_hz;
